[chore] listened to sonarqube

+++ src/main/java/dev/larguma/crawlingmysteries/client/particle/OrbitingStarParticle.java
@@ -12,15 +12,15 @@ public final class OrbitingStarParticle {
       int color, float animationTick, float orbitRadius, int numStars) {
     for (int i = 0; i < numStars; i++) {
       float phaseOffset = (float) (i * Math.PI * 2 / numStars);
-      float angle = animationTick * 0.03f + phaseOffset;
+      float angle = animationTick * 0.03F + phaseOffset;
 
       // Elliptical orbit
       float x = centerX + (float) Math.cos(angle) * orbitRadius;
-      float y = centerY + (float) Math.sin(angle) * orbitRadius * 0.6f;
+      float y = centerY + (float) Math.sin(angle) * orbitRadius * 0.6F;
 
       // alpha based on position (fade when "behind")
-      float depthAlpha = 0.5f + 0.5f * (float) Math.sin(angle);
-      float pulseAlpha = 0.6f + 0.4f * (float) Math.sin(animationTick * 0.1f + phaseOffset);
+      float depthAlpha = 0.5F + 0.5F * (float) Math.sin(angle);
+      float pulseAlpha = 0.6F + 0.4F * (float) Math.sin(animationTick * 0.1F + phaseOffset);
       float alpha = depthAlpha * pulseAlpha;
 
       renderStar(guiGraphics, (int) x, (int) y, 3, color, alpha);
@@ -29,7 +29,7 @@ public final class OrbitingStarParticle {
 
   public static void renderOrbitingStars(GuiGraphics guiGraphics, int centerX, int centerY, int color,
       float animationTick, int slotSize) {
-    renderOrbitingStars(guiGraphics, centerX, centerY, color, animationTick, slotSize / 2 + 4, 3);
+    renderOrbitingStars(guiGraphics, centerX, centerY, color, animationTick, (float) slotSize / 2 + 4, 3);
   }
 
   public static void renderStar(GuiGraphics guiGraphics, int x, int y, int size, int color, float alpha) {
@@ -38,7 +38,7 @@ public final class OrbitingStarParticle {
     guiGraphics.fill(x - size / 2, y - 1, x + size / 2 + 1, y + 2, argb);
     guiGraphics.fill(x - 1, y - size / 2, x + 2, y + size / 2 + 1, argb);
 
-    int brightArgb = RenderUtils.withAlpha(0xFFFFFF, alpha * 0.8f);
+    int brightArgb = RenderUtils.withAlpha(0xFFFFFF, alpha * 0.8F);
     guiGraphics.fill(x, y, x + 1, y + 1, brightArgb);
   }
 }
